backend_new/requirements.txt
fastapi==0.104.1
pandas==2.1.3
numpy==1.26.4
scikit-learn==1.5.2
torch==2.5.1
torchvision==0.20.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
